closes dev-95, dev-96, dev-97

+++ apps/app/api/_lib/cio.ts
@@ -11,7 +11,7 @@ export enum TRANSACTIONAL_EMAILS {
 const client = new APIClient(CIO_API_KEY!);
 const trackClient = new TrackClient(CIO_SITE_ID!, CIO_TRACKING_API_KEY!);
 
-export const deleteUserProfle = ({ userId }) => trackClient.destroy(userId);
+export const deleteUserProfle = ({ userId }: { userId: string }) => trackClient.destroy(userId);
 
 export const sendTransactionalEmail = async ({ messageKey, user, data }: { messageKey: TRANSACTIONAL_EMAILS, user: { id: string; email?: string }, data: any}) => {
   if ( process.env.ENV !== 'production' || !user.email ) { return; }
